removal of classes with an author's name
