+++ src/test/java/school/redrover/NewItem3Test.java
@@ -1,0 +1,63 @@
+package school.redrover;
+
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebElement;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+import school.redrover.runner.BaseTest;
+
+public class NewItem3Test extends BaseTest {
+
+    @Test
+    public void testCreateFreestyleProject() {
+        String expectedResult = "Engineer2";
+
+        getDriver().findElement(By.xpath("//div[@id='tasks']//a[@href='/view/all/newJob']")).click();
+        getDriver().findElement(By.xpath("//input[@id='name']")).sendKeys("Engineer2");
+        getDriver().findElement(By.cssSelector("[value='hudson.model.FreeStyleProject'] + span")).click();
+        getDriver().findElement(By.xpath("//button[@id='ok-button']")).click();
+        getDriver().findElement(By.xpath("//button[@name='Submit']")).click();
+        getDriver().findElement(By.cssSelector("#breadcrumbs > li ")).click();
+
+        String actualResult = getDriver().findElement(By.cssSelector("[href$='Engineer2/']")).getText();
+
+        Assert.assertEquals(actualResult, expectedResult);
+    }
+
+    @Test
+    public void testCreatePipProject() {
+        String expectedPipeline = "Pipeline Engineer";
+        String expectedResult = "Engineer";
+
+        getDriver().findElement(By.xpath("//span[text()='New Item']/../..")).click();
+        WebElement itemName = getDriver().findElement(By.xpath("//input[@name = 'name']"));
+        itemName.click();
+        itemName.sendKeys(expectedResult);
+        getDriver().findElement(By.xpath("//div[@id='items']//li[2]")).click();
+        getDriver().findElement(By.id("ok-button")).click();
+        getDriver().findElement(By.xpath("//button[@formNoValidate='formNoValidate']")).click();
+
+        Assert.assertEquals(getDriver().findElement(By.cssSelector("[class$='headline']")).getText(), expectedPipeline);
+
+        getDriver().findElement(By.cssSelector("#breadcrumbBar > ol > li")).click();
+        String actualResult = getDriver().findElement(By.cssSelector("[href$='Engineer/']")).getText();
+
+        Assert.assertEquals(actualResult, expectedResult);
+    }
+
+    @Test
+    public void testCreateMultiConfigurationProject() {
+
+        String expectedResult = "Engineer3";
+
+        getDriver().findElement(By.cssSelector("[href$='/newJob']")).click();
+        getDriver().findElement(By.cssSelector("input[id='name']")).sendKeys("Engineer3");
+        getDriver().findElement(By.cssSelector("[value$='MatrixProject'] + span")).click();
+        getDriver().findElement(By.cssSelector("#ok-button")).click();
+        getDriver().findElement(By.cssSelector("button[name='Submit']")).click();
+        getDriver().findElement(By.cssSelector("li:nth-child(1) > a")).click();
+
+        WebElement result = getDriver().findElement(By.cssSelector("#projectstatus"));
+        Assert.assertTrue(result.isDisplayed(), "project no display");
+    }
+}
